data.js module was created and 'getRandomAdvert()' was moved to this module
